moved ncurse to sharedconsole
