+++ molotov/sharedconsole.py
@@ -1,74 +1,215 @@
-import sys
+import multiprocess
 import asyncio
 import os
-from queue import Empty
+import signal
 
-import multiprocess
-from molotov.util import cancellable_sleep, printable_error
+from prompt_toolkit import HTML
+from prompt_toolkit.formatted_text import StyleAndTextTuples
+from prompt_toolkit.application import Application
+from prompt_toolkit.formatted_text import to_formatted_text
+from prompt_toolkit.key_binding import KeyBindings
+from prompt_toolkit.layout import (
+    FormattedTextControl,
+    HSplit,
+    Layout,
+    VSplit,
+    Window,
+)
+from prompt_toolkit.layout.controls import UIContent, UIControl
+
+from molotov import __version__
+from molotov.util import printable_error
+
+
+TITLE = HTML(
+    f"<b>Molotov v{__version__}</b> ~ Happy Breaking 🥛🔨 ~ <i>Ctrl+C to abort</i>"
+)
+
+
+class UIControlWithKeys(UIControl):
+    def __init__(self, size=25):
+        super().__init__()
+        self.size = size
+        self._key_bindings = create_key_bindings()
+
+    def is_focusable(self) -> bool:
+        return True  # Make sure that the key bindings work.
+
+    def get_key_bindings(self):
+        return self._key_bindings
+
+
+class TerminalController(UIControlWithKeys):
+    def __init__(self, size=25):
+        super().__init__(size)
+        self.manager = multiprocess.Manager()
+        self.data = self.manager.list()
+
+    def write(self, data):
+        self.data.append(data)
+        if len(self.data) > self.size:
+            self.data[:] = self.data[-self.size :]
+
+    def create_content(self, width, height):
+        items = ["\n"]
+
+        for line in self.data:
+            items.append(line)
+
+        lines = "".join(items).split("\n")
+        items = [to_formatted_text(HTML(line)) for line in lines]
+
+        def get_line(i: int):
+            return items[i]
+
+        return UIContent(get_line=get_line, line_count=len(items), show_cursor=False)
+
+
+def create_key_bindings():
+    kb = KeyBindings()
+
+    @kb.add("c-l")
+    def _clear(event):
+        event.app.renderer.clear()
+
+    @kb.add("c-c")
+    def _interrupt(event):
+        event.app.exit()
+        os.kill(os.getpid(), signal.SIGTERM)
+
+    return kb
+
+
+class RunStatus(UIControlWithKeys):
+    def __init__(self, size=25):
+        super().__init__(size)
+        self.ok = multiprocess.Value("i", 0)
+        self.failed = multiprocess.Value("i", 0)
+        self.worker = multiprocess.Value("i", 0)
+        self.process = multiprocess.Value("i", 0)
+
+    def update(self, results):
+        if "OK" in results:
+            self.ok.value = results["OK"].value
+        if "FAILED" in results:
+            self.failed.value = results["FAILED"].value
+        if "WORKER" in results:
+            self.worker.value = results["WORKER"].value
+        if "PROCESS" in results:
+            self.process.value = results["PROCESS"].value
+
+    def formatted(self):
+        return to_formatted_text(
+            HTML(
+                f'<style fg="green" bg="#cecece">SUCCESS: {self.ok.value} </style>'
+                f'<style fg="red" bg="#cecece"> FAILED: {self.failed.value} </style>'
+                f" WORKERS: {self.worker.value}"
+                f" PROCESSES: {self.process.value}"
+            )
+        )
+
+    def create_content(self, width: int, height: int) -> UIContent:
+        def get_line(i: int) -> StyleAndTextTuples:
+            return self.formatted()
+
+        return UIContent(get_line=get_line, line_count=1, show_cursor=False)
+
+
+class MolotovApp:
+    def __init__(self):
+        self.title = TITLE
+        self.terminal = TerminalController()
+        self.status = RunStatus()
+        self.errors = TerminalController()
+        self.key_bindings = create_key_bindings()
+
+    async def start(self):
+        title_toolbar = Window(
+            FormattedTextControl(lambda: self.title),
+            height=1,
+            style="class:title",
+        )
+
+        bottom_toolbar = Window(
+            content=self.status,
+            style="class:bottom-toolbar",
+            height=1,
+        )
+
+        terminal = Window(content=self.terminal, height=self.terminal.size + 2)
+        errors = Window(content=self.errors, height=self.errors.size + 2)
+
+        self.app = Application(
+            min_redraw_interval=0.05,
+            layout=Layout(
+                HSplit(
+                    [
+                        title_toolbar,
+                        VSplit([terminal, Window(width=4, char=" || "), errors]),
+                        bottom_toolbar,
+                    ]
+                )
+            ),
+            style=None,
+            key_bindings=self.key_bindings,
+            refresh_interval=0.3,
+            color_depth=None,
+            output=None,
+            input=None,
+            erase_when_done=True,
+        )
+
+        self.task = asyncio.ensure_future(self.app.run_async())
+
+    async def stop(self):
+        try:
+            self.app.exit()
+        except Exception:
+            pass
 
 
 class SharedConsole(object):
-    """Multi-process compatible stdout console."""
-
-    def __init__(self, interval=0.1, max_lines_displayed=20, stream=sys.stdout):
-        self._stream = multiprocess.Queue()
+    def __init__(self, interval=0.1, max_lines_displayed=20, stream=None):
         self._interval = interval
         self._stop = True
         self._creator = os.getpid()
         self._stop = False
         self._max_lines_displayed = max_lines_displayed
-        self.stream = stream
+        self.ui = MolotovApp()
+        self.terminal = self.ui.terminal
+        self.errors = self.ui.errors
+        self.status = self.ui.status
+        self.started = False
+
+    async def start(self):
+        await self.ui.start()
+        self.started = True
 
     async def stop(self):
-        self._stop = True
-        while True:
-            try:
-                self.stream.write(self._stream.get_nowait())
-            except Empty:
-                break
-        self.stream.flush()
+        await self.ui.stop()
+        self.started = False
 
-    async def flush(self):
-        self.stream.flush()
-        await asyncio.sleep(0)
+    def print_results(self, results):
+        self.status.update(results)
 
-    async def display(self):
+    def print(self, line):
+        line += "\n"
         if os.getpid() != self._creator:
-            return
-
-        while not self._stop:
-            lines_displayed = 0
-            while True:
-                try:
-                    line = self._stream.get_nowait()
-                    self.stream.write(line)
-                    lines_displayed += 1
-                except Empty:
-                    break
-                if self._stop or lines_displayed > self._max_lines_displayed:
-                    break
-                else:
-                    await asyncio.sleep(0)
-            self.stream.flush()
-            if not self._stop:
-                await cancellable_sleep(self._interval)
-
-    def print(self, line, end="\n"):
-        if os.getpid() != self._creator:
-            line = "[%d] %s" % (os.getpid(), line)
-        line += end
-        self._stream.put_nowait(line)
+            line = f'<style fg="#cecece">[P:{os.getpid()}]</style> {line}'
+        self.terminal.write(line)
 
     def print_error(self, error, tb=None):
         for line in printable_error(error, tb):
-            self.print(line)
+            line = f'<style fg="gray">{line}</style>' + "\n"
+            self.errors.write(line)
+        self.errors.write("\n")
 
     def print_block(self, start, callable, end="OK"):
         if os.getpid() != self._creator:
-            prefix = "[%d] " % os.getpid()
+            prefix = f'<style fg="#cecece">[P:{os.getpid()}]</style>'
         else:
             prefix = ""
-        self._stream.put(prefix + start + "...\n")
+        self.terminal.write(prefix + start + "...\n")
         res = callable()
-        self._stream.put(prefix + "OK\n")
+        self.terminal.write(prefix + "OK\n")
         return res
