+++ molotov/util.py
@@ -7,6 +7,8 @@ import os
 import asyncio
 import time
 import threading
+from collections import defaultdict, MutableSequence
+from contextlib import suppress
 
 from aiohttp import ClientSession, __version__
 
@@ -225,3 +227,43 @@ def printable_error(error, tb=None):
     for line in printed.readlines():
         printable.append(line.rstrip("\n"))
     return printable
+
+
+class SharedTasks(MutableSequence):
+    """Manages tasks lifecycles across processes."""
+
+    def __init__(self):
+        self._tasks = defaultdict(list)
+
+    def _get_tasks(self):
+        return self._tasks[os.getpid()]
+
+    def __len__(self):
+        return len(self._get_tasks())
+
+    def __getitem__(self, i):
+        return self._get_tasks()[i]
+
+    def __delitem__(self, i):
+        del self._get_tasks()[i]
+
+    def __setitem__(self, i, v):
+        self._get_tasks()[i] = v
+
+    def insert(self, i, v):
+        return self._get_tasks().insert(i, v)
+
+    def __str__(self):
+        return str(self._get_tasks())
+
+    def cancel_all(self):
+        cancellable_sleep.cancel_all()
+        for task in reversed(self._get_tasks()):
+            with suppress(asyncio.CancelledError):
+                task.cancel()
+        for task in self._get_tasks():
+            del task
+        self.reset_tasks()
+
+    def reset_tasks(self):
+        self._tasks[os.getpid()][:] = []
